requirements.txt
tqdm
numpy
pandas
tensorflow
scikit-learn
biopython
matplotlib
torch
biopython